Sposta le tabelle achievement da constants.cpp in achievements_data.
Riduce il peso di constants.cpp e allinea il pattern usato per altri dati di dominio (es. procarea_exp), mantenendo le dichiarazioni via constants.hpp.
